+++ components/Navigation/MenuButton.tsx
@@ -1,0 +1,49 @@
+import React from 'react'
+
+import {motion} from 'framer-motion'
+
+
+const Path = props => (
+    <motion.path
+      //fill="transparent"
+      strokeWidth="3"
+      stroke="#189EFE"
+      strokeLinecap="round"
+      {...props}
+    />
+  );
+
+function MenuButton ({open, setOpen}) {
+    return (
+        <motion.button 
+            animate={open ? "open" : "closed"}
+            onClick={setOpen}
+            className='bg-white px-6 py-4 rounded-md shadow-sm z-50'
+        >
+            <svg width="23" height="23" fill="#fff"  viewBox="0 0 23 23">
+                <Path
+                    variants={{
+                        closed: { d: "M 2 2.5 L 20 2.5" },
+                        open: { d: "M 3 16.5 L 17 2.5" }
+                    }}
+                />
+                <Path
+                    d="M 2 9.423 L 20 9.423"
+                    variants={{
+                    closed: { opacity: 1 },
+                    open: { opacity: 0 }
+                    }}
+                    transition={{ duration: 0.1 }}
+                />
+                <Path
+                    variants={{
+                    closed: { d: "M 2 16.346 L 20 16.346" },
+                    open: { d: "M 3 2.5 L 17 16.346" }
+                    }}
+                />
+            </svg>
+        </motion.button>
+    )
+}
+
+export default MenuButton
